refactor: 페이지 컴포넌트는 얇은 셸 역할을 하고, 모든 UI 로직은 features 디렉토리로 이동
