+++ src/app/(with-footer)/entry/[id]/page.tsx
@@ -1,27 +1,24 @@
 import ActionButtons from "@/features/shared/ui/actionButtons";
-import { LinkButton } from "@/features/shared/ui/button";
-import CreateComments from "@/features/survey/entry/components/comment/createComment";
-import EntryClient from "@/features/survey/entry/components/comment/entryClient";
+import { EntryAside } from "@/features/survey/entry/components/EntryAside";
+import { EntryCommentSection } from "@/features/survey/entry/components/EntryCommentSection";
+import { EntryHeader } from "@/features/survey/entry/components/EntryHeader";
+import { MobileImage } from "@/features/survey/entry/components/MobileImage";
 import SimilarForms from "@/features/survey/entry/components/similarForms";
-import Reaction from "@/features/survey/form/components/reaction";
-
-import { RECRUIT_CATEGORY_LABELS, SURVEY_CATEGORY_LABELS } from "@/constants/category";
-import SVGIcon from "@/features/shared/icons/icons";
-import { formatTextWithLineBreaks } from "@/features/shared/ui/formatTextWithLineBreaks";
-import type { Comment, Form } from "@/types";
 
 import { getServerUid } from "@/lib/firebase/auth/getServerUid";
 import { fetchCommentsServer } from "@/lib/firebase/comment/getCommentsServer";
 import { fetchForm } from "@/lib/firebase/form/getFormServer";
 import { handleAnsweredFormIds } from "@/lib/firebase/user/server/handler/handleAnsweredFormIds";
 import { decrypt } from "@/lib/utils/crypoto";
-import formateDate from "@/lib/utils/formateDate";
-
-import Image from "next/image";
 import { notFound } from "next/navigation";
 
 export const dynamic = "force-dynamic";
 
+/**
+ * 엔트리 상세 페이지 컴포넌트
+ * 폼의 상세 정보, 댓글, 관련 폼들을 표시
+ * @param params - 암호화된 폼 ID를 포함한 라우트 파라미터
+ */
 export default async function Entry({
   params,
 }: {
@@ -71,118 +68,6 @@ export default async function Entry({
       />
 
       <SimilarForms itemId={itemId} category={item.category} />
-    </div>
-  );
-}
-
-function EntryHeader({ item, type }: { item: Form; type: string }) {
-  const CATEGORY_LABELS = type === "surveys" ? SURVEY_CATEGORY_LABELS : RECRUIT_CATEGORY_LABELS;
-  return (
-    <div>
-      <div className="subtitle mb-3 space-x-2 text-gray-400">
-        <span>{type === "surveys" ? "#설문조사" : "#모집공고"}</span>
-        <span>{`#${CATEGORY_LABELS[item.category]}`}</span>
-      </div>
-      <h2 className="title3 line-clamp-2 md:text-xl">{item.title}</h2>
-      <hr className="mt-2 mb-3 w-full max-w-120 border border-green-300" />
-      <div className="mb-4 flex gap-3">
-        <span className="caption truncate text-gray-4">
-          {`${formateDate(item.startDate, true)} ~ ${formateDate(item.endDate, true)}`}
-        </span>
-        <Reaction responsesCount={item.responsesCount} commentsCount={item.commentsCount} />
-      </div>
-      {item.description && <p>{formatTextWithLineBreaks(item.description)}</p>}
-    </div>
-  );
-}
-
-function EntryAside({
-  item,
-  encryptedId,
-  hasAnswered,
-}: {
-  item: Form;
-  encryptedId: string;
-  hasAnswered: boolean;
-}) {
-  return (
-    <aside className="flex w-full gap-3 md:w-60 md:flex-col">
-      {item.img && (
-        <Image
-          src={item.img}
-          width={240}
-          height={150}
-          alt="form 이미지"
-          className="hidden rounded-md border border-gray-300 md:flex"
-        />
-      )}
-      {item.isPublic && (
-        <LinkButton className="w-full bg-green-100 text-green-500" href={`/analyze/${encryptedId}`}>
-          결과보기
-        </LinkButton>
-      )}
-      {item.endDate - Date.now() > 0 &&
-        (hasAnswered ? (
-          <LinkButton
-            className="w-full bg-blue-400 text-white"
-            href={`/edit-response/${encryptedId}`}
-          >
-            수정하기
-          </LinkButton>
-        ) : (
-          <LinkButton className="w-full bg-green-400 text-white" href={`/response/${encryptedId}`}>
-            참여하기
-          </LinkButton>
-        ))}
-    </aside>
-  );
-}
-
-function EntryCommentSection({
-  item,
-  initialComments,
-  lastDocId,
-  initialHasNextPage,
-  totalCount,
-}: {
-  item: Form;
-  initialComments: Comment[];
-  lastDocId: string | null;
-  initialHasNextPage: boolean;
-  totalCount: number;
-}) {
-  return (
-    <>
-      <CreateComments id={item.id} />
-      {initialComments && initialComments.length > 0 ? (
-        <EntryClient
-          item={item}
-          initialComments={initialComments}
-          lastDocId={lastDocId}
-          initialHasNextPage={initialHasNextPage}
-          totalCount={totalCount}
-        />
-      ) : (
-        <section className="flex min-h-52 flex-col items-center justify-center gap-4 rounded-2xl border font-bold opacity-20">
-          <SVGIcon name="BubbleChatIcon" />
-          <p>아직 댓글이 없어요. 첫 댓글을 남겨보세요!</p>
-        </section>
-      )}
-    </>
-  );
-}
-
-function MobileImage({ img }: { img: string }) {
-  if (!img) return null;
-  return (
-    <div className="-translate-x-1/2 relative left-1/2 w-screen overflow-hidden shadow md:hidden">
-      <Image
-        src={img}
-        width={240}
-        height={150}
-        alt="form 이미지"
-        className="max-h-[370px] w-full object-cover"
-      />
     </div>
   );
 }
