+++ src/app/(with-footer)/recruit/page.tsx
@@ -1,6 +1,11 @@
 import ListClient from "@/features/survey/list/components/list";
 import { fetchFormList } from "@/lib/firebase/form/getFormListServer";
 
+/**
+ * 모집공고 목록 페이지 컴포넌트
+ * 모집공고 목록을 표시하며 카테고리 및 정렬 옵션을 지원
+ * @param searchParams - URL 쿼리 파라미터 (카테고리, 정렬 옵션)
+ */
 export default async function ListServerWrapper({
   searchParams,
 }: {
