+++ src/app/@modal/(.)entry/[id]/page.tsx
@@ -1,20 +1,20 @@
-import SVGIcon from "@/features/shared/icons/icons";
-import { LinkButton } from "@/features/shared/ui/button";
-import { formatTextWithLineBreaks } from "@/features/shared/ui/formatTextWithLineBreaks";
 import Modal from "@/features/shared/ui/modal";
-import Reaction from "@/features/survey/form/components/reaction";
+import { CommentsSection } from "@/features/survey/entry/components/modal/CommentsSection";
+import { FormContent } from "@/features/survey/entry/components/modal/FormContent";
+import { ViewDetailButton } from "@/features/survey/entry/components/modal/ViewDetailButton";
 
 import { fetchCommentsServer } from "@/lib/firebase/comment/getCommentsServer";
 import { fetchForm } from "@/lib/firebase/form/getFormServer";
 import { decrypt } from "@/lib/utils/crypoto";
-import formatDate from "@/lib/utils/formateDate";
-import type { Comment, Form } from "@/types";
-
-import Image from "next/image";
 import { notFound } from "next/navigation";
 
 export const dynamic = "force-dynamic";
 
+/**
+ * 엔트리 상세 페이지를 모달로 표시하는 페이지 컴포넌트
+ * Intercepting Routes를 사용하여 클라이언트 네비게이션 시 모달로 표시됨
+ * @param params - 암호화된 폼 ID를 포함한 라우트 파라미터
+ */
 export default async function EntryIntercept({ params }: { params: Promise<{ id: string }> }) {
   const { id: encryptedId } = await params;
   const itemId = await decrypt(encryptedId, process.env.CRYPT_SECRET || "");
@@ -31,83 +31,6 @@ export default async function EntryIntercept({ params }: { params: Promise<{ id:
         <CommentsSection comments={initialComments} />
         <ViewDetailButton encryptedId={encryptedId} />
       </Modal>
-    </div>
-  );
-}
-
-function FormContent({ item }: { item: Form }) {
-  return (
-    <section className="space-y-3">
-      <div className="-translate-x-1/2 relative left-1/2 w-screen overflow-hidden shadow md:hidden">
-        {item.img && (
-          <Image
-            src={item.img}
-            width={240}
-            height={150}
-            alt="form 이미지"
-            className="max-h-[370px] w-full object-cover"
-          />
-        )}
-      </div>
-      <h2 className="title3 line-clamp-2 md:text-xl">{item.title}</h2>
-      <hr className="w-full border border-green-300" />
-      {item.description && <p>{formatTextWithLineBreaks(item.description)}</p>}
-      <div className="flex justify-between">
-        <span className="caption truncate text-gray-4">
-          {`${formatDate(item.startDate, true)} ~ ${formatDate(item.endDate, true)}`}
-        </span>
-        <Reaction responsesCount={item.responsesCount} commentsCount={item.commentsCount} />
-      </div>
-    </section>
-  );
-}
-
-function CommentsSection({ comments }: { comments: Comment[] }) {
-  const hasComments = comments && comments.length > 0;
-
-  return (
-    <section
-      className="relative mt-6 h-72 overflow-hidden"
-      aria-labelledby="comments-section-heading"
-    >
-      <h2 id="comments-section-heading" className="sr-only">
-        댓글 섹션
-      </h2>
-
-      {hasComments ? (
-        <ul className="space-y-3">
-          <div className="absolute bottom-0 h-14 w-full bg-linear-to-t from-muted" />
-          {comments.map((comment) => (
-            <li key={comment.id} className="rounded-xl border border-gray-300 bg-surface px-4 py-3">
-              <span className="font-semibold">{comment.displayName}</span>
-              <span className="caption ml-2">
-                {comment.createdAt && new Date(comment.createdAt).toLocaleString()}
-              </span>
-              <p>{comment.content}</p>
-            </li>
-          ))}
-        </ul>
-      ) : (
-        <div className="flex min-h-52 flex-col items-center justify-center gap-4 rounded-2xl border font-bold opacity-20">
-          <SVGIcon name="BubbleChatIcon" size={70} />
-          <p>아직 댓글이 없어요. 첫 댓글을 남겨보세요!</p>
-        </div>
-      )}
-    </section>
-  );
-}
-
-function ViewDetailButton({ encryptedId }: { encryptedId: string }) {
-  return (
-    <div className="text-center">
-      <LinkButton
-        className="bg-green-100 text-green-500"
-        href={`/entry/${encryptedId}`}
-        target="_blank"
-        rel="noopener noreferrer"
-      >
-        자세히 보기
-      </LinkButton>
     </div>
   );
 }
