+++ src/app/auth/sign-in/page.tsx
@@ -1,7 +1,11 @@
 import Form from "@/features/auth/components/sign-in/form";
 import Link from "next/link";
 
-const SignIn: React.FC = () => {
+/**
+ * 로그인 페이지 컴포넌트
+ * 사용자 로그인 폼 및 관련 링크를 표시
+ */
+export default function SignIn() {
   return (
     <>
       <h2 className="title2 mt-20 text-center" id="login-heading">
@@ -25,6 +29,4 @@ const SignIn: React.FC = () => {
       </div>
     </>
   );
-};
-
-export default SignIn;
+}
