+++ src/app/layout.tsx
@@ -7,6 +7,12 @@ import "./globals.css";
 
 const suit = localFont({ src: "../fonts/SUIT.woff2" });
 
+/**
+ * 루트 레이아웃 컴포넌트
+ * 전역 레이아웃, 헤더, 플로팅 컨트롤, 모달 루트를 제공
+ * @param children - 페이지 콘텐츠
+ * @param modal - 모달 콘텐츠 (Intercepting Routes)
+ */
 export default function RootLayout({
   children,
   modal,
